Move gossip server and http server to internal
