--- pkg/bimodal/bimodal.go
@@ -1,29 +1,0 @@
-package bimodal
-
-//import (
-//	"sync"
-//
-//	"github.com/rstefan1/bimodal-multicast/src/internal/buffer"
-//	"github.com/rstefan1/bimodal-multicast/src/internal/config"
-//	"github.com/rstefan1/bimodal-multicast/src/internal/peer"
-//	"github.com/rstefan1/bimodal-multicast/src/gossipgossip"
-//)
-//
-//var (
-//	// shared buffer for peers
-//	peerBuffer = []peer.Peer{}
-//	// shared buffer for messages
-//	msgBuffer = buffer.NewMessageBuffer()
-//)
-//
-//func BMC() error {
-//	stop := make(chan struct{})
-//	cfg := config.Config{}
-//
-//	if err := httpserver.Start(&peerBuffer, &msgBuffer, stop, cfg); err != nil {
-//		return err
-//	}
-//	gossipgossip.Start(&peerBuffer, &msgBuffer)
-//
-//	return nil
-//}
