--- pkg/bimodal/bimodal_test.go
@@ -1,1 +1,0 @@
-package bimodal
